Start changing db to hold 2 mmaps
One mmap is for reading.
The other is for writing.

The page abstraction is still present, but will go away as the
transaction abstraction is completed.

The transaction abstraction will prevent at all times concurrent reading
and writing from the same page, which can currently ahppen as we are
working behind rust's back wth the two mmaps pointing at the same file.

Signed-off-by: Simo Sorce <simo@samba.org>