Simplify root page handling
There is no real need to have two main pages that point to a root page.
As freepages are realy ever ued by writers, simply have a list of free
pages in the zeroth page that is otherwise mostly empty, and just point
to the current root page for readers.

Writers will always allocte a new root page on writing as they have to
replace the whole tree above the leaf they are manipulating.

Signed-off-by: Simo Sorce <simo@samba.org>