Fix function to get parents
This function was failing to find the rightmost node when multiple
layers of nodes were available in the tree, failing to append keys that
are alphabetically the largest.

Change it to always record on the parents vector every node we visit,
and then trimming bck the vector when additional branches are visited.
This returns the very last branch visited regardless of depth allowing
to return the correct node to add a leaf to when the key is the largest.

Signed-off-by: Simo Sorce <simo@samba.org>