Fix crash when adding on split node boundary
When adding on a boundary of a split node, we need to chose between on
of the two returned halves. Doing a full search could land us on a
neighboring branch that also turns out to have a full parent. This was
causing an InvalidMetadata error afterwards, when we were going to add
the new leaf on a full parent.

Signed-off-by: Simo Sorce <simo@samba.org>